Edit: Moved utility code to from views module to utils module
